Extract `show_region` to `core/view.py`
